mock/requirements.txt
kafka-python==2.0.2
numpy==1.19.5
pandas==1.3.4
httpx==0.20.0
requests==2.26.0